third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-grid-flex-multicol-expected.txt

This is a testharness.js-based test.
[FAIL] inline-flex
  assert_equals: display expected "inline-flex" but got "flex"
[FAIL] inline-grid
  assert_equals: display expected "inline-grid" but got "grid"
Harness: the test ran to completion.
